requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
docker==7.1.0
fastapi==0.116.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.16.0
httptools==0.6.4
idna==3.10
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
Wand==0.6.13
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
